fix: end to end testes
LBSD-474
