+++ client/app/scripts/liveblog-edit/views/settings.html
@@ -26,7 +26,7 @@
                     <a ng-href="" ng-click="settings.generalTab = true">General</a>
                 </li>
                 <li ng-class="{active: !settings.generalTab}" data="blog-settings-users">
-                    <a ng-href="" ng-click="settings.generalTab = false">Users</a>
+                    <a ng-href="" ng-click="settings.generalTab = false" translate>Users</a>
                 </li>
             </ul>
             <div>
